@@ -1,31 +1,28 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.3.10'
-    ext.support_library_version = '28.0.0'
-    ext.dokka_version = '0.9.17'
-    ext.buildtools_version = '28.0.3'
-    repositories {
-        jcenter()
-        google()
-    }
-    dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.1'
-        classpath "org.jetbrains.dokka:dokka-android-gradle-plugin:$dokka_version"
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-
-        classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.4'
-        classpath 'com.github.dcendents:android-maven-gradle-plugin:2.1'
-    }
+  ext.kotlin_version = '1.8.0'
+  ext.support_library_version = '28.0.0'
+  ext.dokka_version = '0.9.17'
+  ext.buildtools_version = '28.0.3'
+  repositories {
+    jcenter()
+    google()
+  }
+  dependencies {
+    classpath 'com.android.tools.build:gradle:8.5.0'
+    //        classpath "org.jetbrains.dokka:dokka-android-gradle-plugin:$dokka_version"
+    classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+  }
 }
 
 allprojects {
-    repositories {
-        jcenter()
-        google()
-    }
+  repositories {
+    jcenter()
+    google()
+  }
 }
 
 task clean(type: Delete) {
-    delete rootProject.buildDir
+  delete rootProject.buildDir
 }
