@@ -1,14 +1,14 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 
-
 android {
-    compileSdkVersion 28
+    compileSdkVersion 34
     buildToolsVersion buildtools_version
+    namespace = "com.arasthel.spannedgridlayoutmanager.sample"
     defaultConfig {
         applicationId "com.arasthel.spannedgridlayoutmanager.sample"
-        minSdkVersion 14
-        targetSdkVersion 28
+        minSdkVersion 26
+        targetSdkVersion 34
         versionCode 1
         versionName "1.0"
     }
@@ -18,12 +18,18 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    kotlinOptions {
+        jvmTarget = "1.8" // Set to a valid JVM target version
+    }
 }
 
 dependencies {
-    implementation "com.android.support:appcompat-v7:$support_library_version"
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation "com.android.support:recyclerview-v7:$support_library_version"
+    implementation 'androidx.core:core:1.13.1'
+    implementation 'androidx.appcompat:appcompat:1.7.0'
+    //  implementation "com.android.support:appcompat-v7:$support_library_version"
+    //  implementation 'com.android.support.constraint:constraint-layout:2.0.4'
+    //  implementation "com.android.support:recyclerview-v7:$support_library_version"
+    implementation "androidx.recyclerview:recyclerview:1.3.2"
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 
     implementation project(':spannedgridlayoutmanager')
