@@ -1,60 +1,61 @@
 apply plugin: 'com.android.library'
 apply plugin: 'kotlin-android'
-apply plugin: 'org.jetbrains.dokka-android'
+//apply plugin: 'org.jetbrains.dokka-android'
 
-task dokkaJavadoc(type: org.jetbrains.dokka.gradle.DokkaAndroidTask) {
-    delete "$rootDir/docs"
+//task dokkaJavadoc(type: org.jetbrains.dokka.gradle.DokkaAndroidTask) {
+//  delete "$rootDir/docs"
+//
+//  outputFormat = 'html'
+//  outputDirectory = "$rootDir/docs"
+//
+//  externalDocumentationLink {
+//    url = new URL("https://developer.android.com/reference/")
+//  }
+//
+//  includeNonPublic = false
+//
+//  // Always recreate documentation
+//  outputs.upToDateWhen { return false }
+//}
 
-    outputFormat = 'html'
-    outputDirectory = "$rootDir/docs"
-
-    externalDocumentationLink {
-        url = new URL("https://developer.android.com/reference/")
-    }
-
-    includeNonPublic = false
-
-    // Always recreate documentation
-    outputs.upToDateWhen { return false }
-}
-
-afterEvaluate {
-    if (project.hasProperty("javadocJar")) {
-        tasks.javadocJar.dependsOn dokkaJavadoc
-    }
-}
+//afterEvaluate {
+//  if (project.hasProperty("javadocJar")) {
+//    tasks.javadocJar.dependsOn dokkaJavadoc
+//  }
+//}
 
 ext {
-    bintrayRepo = 'maven'
-    bintrayName = 'spannedgridlayoutmanager'
-
-    publishedGroupId = 'com.arasthel'
-    libraryName = 'SpannedGridLayoutManager'
-    artifact = 'spannedgridlayoutmanager'
-
-    libraryDescription = 'A layout manager that will resize and reorder views based on a provided SpanSize.'
-
-    siteUrl = 'https://github.com/Arasthel/SpannedGridLayoutManager'
-    gitUrl = 'https://github.com/Arasthel/SpannedGridLayoutManager.git'
-
+    //  bintrayRepo = 'maven'
+    //  bintrayName = 'spannedgridlayoutmanager'
+    //
+    //  publishedGroupId = 'com.arasthel'
+    //  libraryName = 'SpannedGridLayoutManager'
+    //  artifact = 'spannedgridlayoutmanager'
+    //
+    //  libraryDescription = 'A layout manager that will resize and reorder views based on a provided SpanSize.'
+    //
+    //  siteUrl = 'https://github.com/Arasthel/SpannedGridLayoutManager'
+    //  gitUrl = 'https://github.com/Arasthel/SpannedGridLayoutManager.git'
+    //
     libraryVersion = '3.0.2'
-
-    developerId = 'Arasthel'
-    developerName = 'Jorge Martín Epsinosa'
-    developerEmail = 'jorgemartinespinosa@gmail.com'
-
-    licenseName = 'MIT'
-    licenseUrl = 'https://opensource.org/licenses/MIT'
-    allLicenses = ["MIT"]
+    //
+    //  developerId = 'Arasthel'
+    //  developerName = 'Jorge Martín Epsinosa'
+    //  developerEmail = 'jorgemartinespinosa@gmail.com'
+    //
+    //  licenseName = 'MIT'
+    //  licenseUrl = 'https://opensource.org/licenses/MIT'
+    //  allLicenses = ["MIT"]
 }
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 34
     buildToolsVersion buildtools_version
+    namespace = "com.arasthel.spannedgridlayoutmanager"
 
     defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 28
+        minSdkVersion 26
+        targetSdkVersion 34
         versionCode 1
         versionName libraryVersion
     }
@@ -64,6 +65,9 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    kotlinOptions {
+        jvmTarget = "1.8" // Set to a valid JVM target version
+    }
 }
 
 tasks.withType(Javadoc) {
@@ -71,10 +75,11 @@ tasks.withType(Javadoc) {
 }
 
 dependencies {
-    implementation "com.android.support:recyclerview-v7:$support_library_version"
+    //  implementation "com.android.support:recyclerview-v7:$support_library_version"
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+    implementation "androidx.recyclerview:recyclerview:1.3.2"
 }
 
 // Bintray Upload
-apply from: '../gradle/tools/bintrayv1.gradle'
-apply from: '../gradle/tools/installv1.gradle'
+//apply from: '../gradle/tools/bintrayv1.gradle'
+//apply from: '../gradle/tools/installv1.gradle'
