@@ -9,7 +9,7 @@ import android.widget.TextView
 /**
  * Created by Jorge Martín on 24/5/17.
  */
-class GridItemView(context: Context?) : FrameLayout(context) {
+class GridItemView(context: Context) : FrameLayout(context) {
 
     private var title: TextView
 
